requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
langchain-text-splitters
langchain-pinecone
pinecone-client
streamlit
python-dotenv
openai
tiktoken